+++ src/main/java/dmo/fs/spa/utils/SpaUtil.java
@@ -152,4 +152,8 @@ public class SpaUtil {
         
         return spaLogin;
       }
+
+      public static SpaLogin createSpaLogin() {
+		return new SpaLoginImpl();
+	}
 }